feat: add gas estimation for cross-chain transactions
