+++ .env.example
@@ -1,2 +1,3 @@
 SUI_PRIVATE_KEY=
 XRPL_SEED=
+ENVIRONMENT=testnet
