+++ common/gasEstimation.ts
@@ -1,0 +1,32 @@
+import { AxelarQueryAPI, type Environment } from "@axelar-network/axelarjs-sdk";
+
+export type HopParams = {
+  sourceChain: string;
+  destinationChain: string;
+  gasLimit: string;
+};
+
+const environment = process.env.ENVIRONMENT || "testnet";
+
+export async function calculateEstimatedFee(
+  sourceChain: string,
+  destinationChain: string,
+): Promise<string> {
+  const sdk = new AxelarQueryAPI({
+    environment: environment as Environment,
+  });
+
+  const hopParams: HopParams[] = [
+    {
+      sourceChain: sourceChain,
+      destinationChain: "axelar",
+      gasLimit: "400000",
+    },
+    {
+      sourceChain: "axelar",
+      destinationChain: destinationChain,
+      gasLimit: "1100000",
+    },
+  ];
+  return (await sdk.estimateMultihopFee(hopParams)) as string;
+}
