+++ sui-to-evm.ts
@@ -6,6 +6,7 @@ import { AxelarQueryAPI, Environment } from "@axelar-network/axelarjs-sdk";
 import { formatUnits, parseUnits } from "ethers";
 import SuiTypedContracts from "@axelarjs/sui";
 import { getItsCoin } from "./sui/coin";
+import { calculateEstimatedFee } from "./common/gasEstimation";
 
 // --- Constants ---
 const DESTINATION_CHAIN = "ethereum-sepolia";
@@ -24,33 +25,6 @@ const ITS_TOKEN_ID =
   "0x42e69c5a9903ba193f3e9214d41b1ad495faace3ca712fb0c9d0c44cc4d31a0c";
 const ITS_TOKEN_TYPE = `${TOKEN_ADDRESS}::${TOKEN_SYMBOL.toLowerCase()}::${TOKEN_SYMBOL.toUpperCase()}`;
 const CLOCK_PACKAGE_ID = "0x6";
-
-type HopParams = {
-  sourceChain: string;
-  destinationChain: string;
-  gasLimit: string;
-};
-
-// --- Helper Functions ---
-async function calculateEstimatedFee(
-  sdk: AxelarQueryAPI,
-  sourceChain: string,
-  destinationChain: string,
-): Promise<string> {
-  const hopParams: HopParams[] = [
-    {
-      sourceChain: sourceChain,
-      destinationChain: "axelar",
-      gasLimit: "400000",
-    },
-    {
-      sourceChain: "axelar",
-      destinationChain: destinationChain,
-      gasLimit: "1100000",
-    },
-  ];
-  return (await sdk.estimateMultihopFee(hopParams)) as string;
-}
 
 // --- Main Execution ---
 (async () => {
@@ -76,11 +50,7 @@ async function calculateEstimatedFee(
     gasService: contracts.GasService.objects.GasService,
   };
 
-  const sdk = new AxelarQueryAPI({
-    environment: ENVIRONMENT,
-  });
-
-  const fee = await calculateEstimatedFee(sdk, "sui", DESTINATION_CHAIN);
+  const fee = await calculateEstimatedFee("sui", DESTINATION_CHAIN);
   console.log("Estimated Fee:", `${formatUnits(fee, 9)} SUI`);
 
   console.log(
